move ble native mock class to separate file (should possibly be placed somewhere else)
